% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{approxSilhouette}
\alias{approxSilhouette}
\title{Internal Function 2}
\usage{
approxSilhouette(x, clusters)
}
\arguments{
\item{data}{A matrix of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.}

\item{clustering}{cluster labels}
}
\value{
A data frame of silhouette coefficients for each partition.
}
\description{
This function outputs silhouette coefficients.
}
\keyword{internal}
