% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{BPV}
\alias{BPV}
\title{BPV data}
\format{
A data frame of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.

\describe{
\item{channel1}{fluorescence intensities of color 1}
\item{channel2}{fluorescence intensities of color 2}
\item{channel3}{fluorescence intensities of color 3}
}
}
\usage{
BPV
}
\description{
A 3-color dPCR data of bovine papilloma virus assay
}
\examples{
data(BPV)
}
\keyword{datasets}
