% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{CA}
\alias{CA}
\title{CA data}
\format{
A data frame of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.
data is not orthogonal.
\describe{
\item{channel1}{fluorescence intensities of color 1}
\item{channel2}{fluorescence intensities of color 2}
}
}
\usage{
CA
}
\description{
2-color competitive assay of competition BRAF V600E assay with 1\% mutant
}
\examples{
data(CA)
}
\keyword{datasets}
