% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{cluster_selection}
\alias{cluster_selection}
\title{Internal Function 11}
\usage{
cluster_selection(cluster_num)
}
\arguments{
\item{cluster_num}{The expected maximum number of clusters}
}
\value{
A matrix of all combinations of primary targets
}
\description{
This function outputs all combinations of primary targets
}
\keyword{internal}
