% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{CNV6plex}
\alias{CNV6plex}
\title{CNV 6-plex data}
\format{
A data frame of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.

\describe{
\item{channel1}{fluorescence intensities of color 1}
\item{channel2}{fluorescence intensities of color 2}
\item{channel3}{fluorescence intensities of color 3}
\item{channel4}{fluorescence intensities of color 4}
\item{channel5}{fluorescence intensities of color 5}
\item{channel6}{fluorescence intensities of color 6}
}
}
\usage{
CNV6plex
}
\description{
CNV 6-plex universal probes
}
\examples{
data(CNV6plex)
}
\keyword{datasets}
