% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{combined_vectors}
\alias{combined_vectors}
\title{Internal Function 4}
\usage{
combined_vectors(coefs, mus, cluster_num, dim_data)
}
\arguments{
\item{coefs}{coefs The coefficients to adjust for the expected cluster centers. The default is 1 which can be used for common assay designs and has
to be modified for special assays such as competing assays.}

\item{mus}{The cluster centers of primary targets}

\item{cluster_num}{The expected maximum number of clusters.}

\item{dim_data}{dimension of the dataset}
}
\value{
A list of vectors and weights
}
\description{
This function outputs vectors and weights that will be used in EM algorithm
}
\keyword{internal}
