% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{compute_tmp_matrix}
\alias{compute_tmp_matrix}
\title{Internal Function 6}
\usage{
compute_tmp_matrix(g, k, cluster_num, mg, log_pih, mug_t, muh_t, covh, covg)
}
\arguments{
\item{g}{cluster index}

\item{k}{cluster index}

\item{cluster_num}{The expected maximum number of clusters}

\item{mg}{cluster sizes of base clustering result}

\item{log_pih}{log pih (the probability of cluster g belonging at level l+1 to cluster h at level l)}

\item{mug_t}{the transposed matrix of cluster centers at level l+1}

\item{muh_t}{the transposed matrix of cluster centers at level l}

\item{covh}{the covariance matrix of clusters at level l}

\item{covg}{the covariance matrix of clusters at level l+1}
}
\value{
A vector of intermediate values for zi calculation in estep function
}
\description{
This function compute the necessary elements for estep function
}
\keyword{internal}
