% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{estep}
\alias{estep}
\title{Internal Function 7}
\usage{
estep(g_clusternum, cluster_num, pih, muh, covh, mg, mug, covg)
}
\arguments{
\item{g_clusternum}{cluster labels from base clustering}

\item{cluster_num}{The expected maximum number of clusters}

\item{pih}{the probability of cluster g belonging at level l+1 to cluster h at level l}

\item{muh}{the matrix of cluster centers at level l}

\item{covh}{the covariance matrix of clusters at level l}

\item{mg}{cluster sizes of base clustering result}

\item{mug}{the matrix of cluster centers at level l+1}

\item{covg}{the covariance matrix of clusters at level l+1}
}
\value{
zi for estep in EM algorithm
}
\description{
This function calculates zi in E-step of EM algorithm
}
\keyword{internal}
