% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{fp_search}
\alias{fp_search}
\title{Internal Function 3}
\usage{
fp_search(data, cluster_num = 16)
}
\arguments{
\item{data}{A matrix of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.}

\item{cluster_num}{The expected maximum number of clusters}
}
\value{
A vector containing the optimal parameters found by the algorithm
}
\description{
This function optimizes parameters of flowPeaks
}
\keyword{internal}
