% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{GMM_init}
\alias{GMM_init}
\title{Internal Function 5}
\usage{
GMM_init(data, cluster_num, base_clust, coefs)
}
\arguments{
\item{data}{A matrix or data frame of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.}

\item{cluster_num}{The expected maximum number of clusters}

\item{base_clust}{The results of base clustering}

\item{coefs}{The coefficients to adjust for the expected cluster centers. The default is 1 which can be used for common assay designs and has
to be modified for special assays such as competing assays.}
}
\value{
A list of initial parameters for the EM algorithm
}
\description{
This function intialize the parameters for the main clustering function
}
\keyword{internal}
