% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{LR}
\alias{LR}
\title{LR data}
\format{
A data frame of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.
barely separable on x-axis.
\describe{
\item{channel1}{fluorescence intensities of color 1}
\item{channel2}{fluorescence intensities of color 2}
}
}
\usage{
data(LR)
}
\description{
A low-resolution 2-color dPCR data of development of genotyping assays for plants various
}
\examples{
data(LR)
head(LR)
}
\keyword{datasets}
