% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytect_plot.R
\name{polytect_plot}
\alias{polytect_plot}
\title{plotting function}
\usage{
polytect_plot(df_data)
}
\arguments{
\item{df_data}{A data frame containing partition fluorescence intensities and corresponding cluster label. This can be the output
of \code{polytect_merge} or any data frame containing the above information.}
}
\value{
2-d plots.
}
\description{
plotting function
}
\examples{
data(HR)
df_data<-polytect_merge(HR,4)
polytect_plot(df_data)
}
