% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{silhouette_coef}
\alias{silhouette_coef}
\title{Internal Function 1}
\usage{
silhouette_coef(
  data,
  clustering,
  plot = FALSE,
  plot_name = "orig",
  sim = "orig"
)
}
\arguments{
\item{data}{A data frame containing standardized partition fluorescence intensities and corresponding cluster label.}

\item{clustering}{cluster labels}
}
\value{
A list of silhouette coefficients for each partition and the mean silhouette coefficients for each cluster.
}
\description{
This function outputs silhouette coefficients.
}
\keyword{internal}
