% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_cal.R
\name{conc_cal}
\alias{conc_cal}
\title{concentration calculation function}
\usage{
conc_cal(df_data, sampvol = 0.91, volmix = 20, voltemp = 20, type = "2color")
}
\arguments{
\item{df_data}{A data frame containing partition fluorescence intensities and corresponding cluster label. This can be the output
of \code{polytect_merge} or any data frame containing the above information.}

\item{sampvol}{The sample volume in microliters (µL)}

\item{volmix}{The volume of the mixture}

\item{voltemp}{The volume of the template}

\item{type}{The assay design, including the number of channels and targets. \code{type}=c("2color",
"2colorHO","3color","4color"). "2color" is chosen when there are 2 colors and 2 targets. "2colorHO" means higher-order 2-color data (2 color and 3 targets). "3color" means
3-color and 3-target. "4-color" is chosen when there are 4 colors.}
}
\value{
a data frame of the summary of cluster centers, cluster sizes and cluster silhouette coefficients.
}
\description{
concentration calculation function
}
\examples{
data(HR)
df_data<-polytect_clust(HR,4)
conc_cal(df_data)
}
