% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{HR}
\alias{HR}
\title{HR data}
\format{
A data frame of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.
good separation but some crosstalk; RPP30 genomic DNA assay
\describe{
\item{channel1}{channel 1}
\item{channel2}{channel 2}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/33992770/}
}
\usage{
HR
}
\description{
A high-resolution 2-d dPCR data
}
\keyword{datasets}
