% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{LR}
\alias{LR}
\title{LR data}
\format{
A data frame of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.
barely separable on x-axis; development of genotyping assays for plants various; primers/probes and conditions for each target are evaluated.
\describe{
\item{channel1}{channel 1}
\item{channel2}{channel 2}
}
}
\usage{
LR
}
\description{
A low-resolution 2-d dPCR data
}
\keyword{datasets}
