% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytect_clust.R
\name{polytect_clust}
\alias{polytect_clust}
\title{General function for flowpeaks and merging}
\usage{
polytect_clust(
  data,
  cluster_num,
  type = "2color",
  fp_par = "default",
  fp_optim = c(0.1, 1, 1.5),
  lambdas = rep(2, 12),
  coefs = rep(1, 4)
)
}
\arguments{
\item{data}{A matrix of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.}

\item{cluster_num}{The expected maximum number of clusters.}

\item{type}{The assay design, including the number of channels and targets. \code{type}=c("2color",
"2colorHO","3color","4color"). "2color" is chosen when there are 2 colors and 2 targets. "2colorHO" means higher-order 2-color data (2 color and 3 targets). "3color" means
3-color and 3-target. "4-color" is chosen when there are 4 colors.}

\item{fp_par}{The parameters for flowPeaks. \code{fp_par}=c("default","manual","auto"). When "default" is chosen, the default parameters of
flowPeaks will be used. With "manual", you have to fill in \code{fp_optim}.}

\item{fp_optim}{The paramters for flowPeaks that users have to fill in manually when \code{fp_par} is set at "manual".}

\item{lambdas}{The penalty terms for the deviation from the expected cluster centers. Higher \code{lambdas} penalizes the deviation more.}

\item{coefs}{The coefficients to adjust for the expected cluster centers. The default is 1 which can be used for common assay designs and has
to be modified for special assays such as competing assays.}
}
\value{
A data frame containing the original fluorescence intensity and the cluster labels.
}
\description{
General function for flowpeaks and merging
}
\examples{
data(HR)
polytect_clust(HR, 4)
}
