% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytect_merge.R
\name{polytect_merge}
\alias{polytect_merge}
\title{Function for merging}
\usage{
polytect_merge(
  data,
  cluster_num,
  base_clust,
  lambdas = rep(2, 64 - log2(64)),
  coefs = rep(1, 6)
)
}
\arguments{
\item{data}{A matrix of fluorescence intensities in each channel. Each row represents each partitions, and each column each channel.}

\item{cluster_num}{The expected maximum number of clusters.}

\item{base_clust}{A list that contains partition labels given by initial clustering.}

\item{type}{The assay design, including the number of channels and targets. \code{type}=c("2color",
"2colorHO","3color","4color"). "2color" is chosen when there are 2 colors and 2 targets. "2colorHO" means higher-order 2-color data (2 color and 3 targets). "3color" means
3-color and 3-target. "4-color" is chosen when there are 4 colors.}
}
\value{
A data frame containing the original fluorescence intensity and the cluster labels.
}
\description{
This function takes the clustering result as input. Users can first perform any clustering algorithm, then use this function. It
will return a data frame of fluorescence intensities and partition labels.
}
\examples{
data(HR)
dist_matrix <- dist(HR)
hc <- hclust(dist_matrix, method = "ward.D2")
hc_clusters <- cutree(hc, k = 6)
polytect_merge(HR, 4, hc_clusters)
}
