% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sil_plot.R
\name{sil_plot}
\alias{sil_plot}
\title{plotting function}
\usage{
sil_plot(df_data)
}
\arguments{
\item{df_data}{A data frame containing partition fluorescence intensities and corresponding cluster label. This can be the output
of \code{polytect_merge} or any data frame containing the above information.}
}
\value{
plot of silhouette coefficients for each cluster.
}
\description{
plotting function
}
\examples{
data(HR)
df_data<-polytect_merge(HR,4)
sil_plot(df_data)
}
